Fix JsonSyntaxException crash in LogModel
Fixes https://gitlab.com/Nulide/findmydevice/-/issues/271

For now, let's only catch this for the LogRepository.
Arbitrary string may end up in the log, maybe this is what causing the syntax exception
(even though it shouldn't, assuming that Gson.toJson is correct).

The other repositories that use Gson (Settings, Allowlist) store much more simple data.
Also, silently clearing logs is fine, but silently resetting settings is not.

Therefore, only apply this fix to the LogRepository for now.
If we later get reports of the JsonSyntaxException being thrown in other repos, we can reconsider.